backend-fastapi/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
authlib
httpx
itsdangerous
python-dotenv
